feat: Refactor context utils.
