+++ components/ADempiere/FieldDefinition/FieldWarehouseLocator/index.vue
@@ -66,7 +66,7 @@ import { WAREHOUSE } from '@/utils/ADempiere/constants/systemColumns'
 
 // Utils and Helper Methods
 import { isEmptyValue } from '@/utils/ADempiere/valueUtils'
-import { getContextAttributes } from '@/utils/ADempiere/contextUtils'
+import { getContextAttributes } from '@/utils/ADempiere/contextUtils/contextAttributes'
 
 export default {
   name: 'FieldWarehouseLocator',
