+++ components/ADempiere/FieldDefinition/FieldWarehouseLocator/useLocatorWarehouse.js
@@ -1,6 +1,6 @@
 /**
  * ADempiere-Vue (Frontend) for ADempiere ERP & CRM Smart Business Solution
- * Copyright (C) 2017-Present E.R.P. Consultores y Asociados, C.A. www.erpya.com
+ * Copyright (C) 2018-Present E.R.P. Consultores y Asociados, C.A. www.erpya.com
  * Contributor(s): Edwin Betancourt EdwinBetanc0urt@outlook.com https://github.com/EdwinBetanc0urt
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -16,9 +16,9 @@
  * along with this program. If not, see <https://www.gnu.org/licenses/>.
  */
 
-import { computed } from '@vue/composition-api'
-
 import store from '@/store'
+
+import { computed } from '@vue/composition-api'
 
 // Constants
 import {
@@ -28,7 +28,7 @@ import { WAREHOUSE } from '@/utils/ADempiere/constants/systemColumns'
 
 // Utils and Helper Methods
 import { isEmptyValue } from '@/utils/ADempiere/valueUtils'
-import { getContextAttributes } from '@/utils/ADempiere/contextUtils'
+import { getContextAttributes } from '@/utils/ADempiere/contextUtils/contextAttributes'
 import useDisplayedColumn from '@theme/components/ADempiere/FieldDefinition/useDisplayedColumn.js'
 
 export default ({
