+++ components/ADempiere/TabManager/tabChild.vue
@@ -1,6 +1,6 @@
 <!--
  ADempiere-Vue (Frontend) for ADempiere ERP & CRM Smart Business Solution
- Copyright (C) 2017-Present E.R.P. Consultores y Asociados, C.A. www.erpya.com
+ Copyright (C) 2018-Present E.R.P. Consultores y Asociados, C.A. www.erpya.com
  Contributor(s): Edwin Betancourt EdwinBetanc0urt@outlook.com https://github.com/EdwinBetanc0urt
  This program is free software: you can redistribute it and/or modify
  it under the terms of the GNU General Public License as published by
@@ -137,7 +137,7 @@ import { UUID } from '@/utils/ADempiere/constants/systemColumns.js'
 // utils and helper methods
 import { isEmptyValue } from '@/utils/ADempiere/valueUtils.js'
 import { isDisplayedTab } from '@/utils/ADempiere/dictionary/window'
-import { getContextAttributes, generateContextKey } from '@/utils/ADempiere/contextUtils'
+import { getContextAttributes, generateContextKey } from '@/utils/ADempiere/contextUtils/contextAttributes'
 
 export default defineComponent({
   name: 'TabManagerChild',
